requirements.txt
colorlog==6.9.0
pandas==2.2.3
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36